test: move bootstrap audit tests to -upgrade.js
less relevant to -integration.js
