+++ packages/vats/test/bootstrapTests/test-vaults-integration.js
@@ -4,10 +4,10 @@
  */
 import { test as anyTest } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 
-import { Fail, NonNullish } from '@agoric/assert';
+import { Fail } from '@agoric/assert';
 import { Offers } from '@agoric/inter-protocol/src/clientSupport.js';
 import { eventLoopIteration } from '@agoric/internal/src/testing-utils.js';
-import { Far, makeMarshal } from '@endo/marshal';
+import { makeMarshal } from '@endo/marshal';
 import {
   makeAgoricNamesRemotesFromFakeStorage,
   slotToBoardRemote,
@@ -67,95 +67,6 @@ test.before(async t => {
 });
 test.after.always(t => {
   t.context.shutdown && t.context.shutdown();
-});
-
-test('audit bootstrap exports', async t => {
-  const expected = {
-    maxExports: 5,
-    maxNonDurable: 5,
-    ifaces: {
-      // in bridgeCoreEval()
-      coreHandler: true,
-      // in bridgeProvisioner()
-      provisioningHandler: true,
-      'prioritySenders manager': true,
-      // TODO? move to provisioning vat?
-      clientCreator: true,
-    },
-  };
-
-  const { controller } = t.context;
-  const kState = controller.dump();
-
-  const myVatID = 'v1';
-
-  const myPromises = kState.promises.filter(
-    // @ts-expect-error kernel.dump() .promises type is wrong
-    p => p.decider === myVatID,
-  );
-  t.true(myPromises.length <= 1, 'bootstrap is the decider of only its return');
-
-  const myExports = kState.kernelTable.filter(
-    o => o[1] === myVatID && o[2].startsWith('o+'),
-  );
-  const v1VatTable =
-    kState.vatTables.find(vt => vt.vatID === myVatID) || assert.fail();
-  const { transcript } = v1VatTable.state;
-
-  const oids = new Set(myExports.map(o => o[2]));
-  const oidsDurable = [...oids].filter(o => o.startsWith('o+d'));
-  t.log(
-    'bootstrap exports:',
-    oidsDurable.length,
-    'durable',
-    oids.size - oidsDurable.length,
-    'non-durable',
-    oids.size,
-    'total',
-  );
-  t.true(oids.size <= expected.maxExports, 'too many exports');
-  t.true(
-    oids.size - oidsDurable.length <= expected.maxNonDurable,
-    'too many non-durable',
-  );
-
-  // Map oid to iface by poring over transcript syscalls
-  const toIface = new Map();
-  const anObj = Far('obj', {});
-  const aPromise = harden(new Promise(() => {}));
-  const saveBootstrapIface = (slot, iface) => {
-    if (slot.startsWith('p')) return aPromise;
-    if (oids.has(slot)) {
-      toIface.set(slot, iface);
-    }
-    return anObj;
-  };
-  const m = makeMarshal(undefined, saveBootstrapIface);
-  oids.forEach(oid => {
-    for (const [_ix, ev] of transcript) {
-      for (const sc of ev.sc) {
-        if (sc.s[0] === 'send') {
-          const { methargs } = sc.s[2];
-          if (!methargs.slots.includes(oid)) continue;
-          m.fromCapData(methargs);
-          return;
-        } else if (sc.s[0] === 'resolve') {
-          for (const res of sc.s[1]) {
-            const capdata = res[2];
-            if (!capdata.slots.includes(oid)) continue;
-            m.fromCapData(capdata);
-            return;
-          }
-        }
-      }
-    }
-  });
-
-  const exportedInterfaces = Object.fromEntries(
-    [...toIface.values()].map(iface => [iface.replace(/^Alleged: /, ''), true]),
-  );
-
-  t.deepEqual(exportedInterfaces, expected.ifaces, 'expected interfaces');
 });
 
 test('metrics path', async t => {
@@ -448,144 +359,4 @@ test('propose change to auction governance param', async t => {
   const changes = lastQuestion?.issue?.spec?.changes;
   t.log('check Economic_Committee.latestQuestion against proposal');
   t.like(changes, { StartFrequency: { relValue: 300n } });
-});
-
-test('upgrade facet for all contracts, vats are saved durably', async t => {
-  const {
-    controller,
-    runUtils: { EV },
-    swingStore,
-  } = t.context;
-  const kState = controller.dump();
-  const { kernelTable, vatTables } = kState;
-
-  t.log('find contracts, vats known to SwingSet');
-  // see comment atop kernelKeeper.js for schema
-  const vatOptions = vatID =>
-    JSON.parse(
-      NonNullish(
-        swingStore.kernelStorage.kvStore.get(`${vatID}.options`),
-        vatID,
-      ),
-    );
-  const vatNames = Object.fromEntries(
-    vatTables.map(vt => [vt.vatID, vatOptions(vt.vatID).name]),
-  );
-  const vatNamed = name =>
-    Object.keys(vatNames).find(vatID => vatNames[vatID] === name) ||
-    assert.fail();
-  const zoeVat = vatNamed('zoe');
-
-  t.log('discharge obligations by finding upgrade powers');
-  const todo = Object.fromEntries(Object.entries(vatNames));
-  // vats created by swingset before bootstrap starts
-  const swingsetVats = ['bootstrap', 'comms', 'vatAdmin', 'vattp', 'timer'].map(
-    vatNamed,
-  );
-  swingsetVats.forEach(vatID => {
-    delete todo[vatID];
-  });
-
-  // TODO? test that powerStore is in baggage?
-  // by inspection, we see that it is:
-  // [ "v1.vs.vc.1.sBootstrap Powers", "{\"body\":\"#\\\"$0.Alleged: mapStore\\\"\",\"slots\":[\"o+d6/5\"]}",]
-
-  /** @type {MapStore} */
-  const powerStore = await EV.vat('bootstrap').consumeItem('powerStore');
-
-  /** @type {(n: string) => Promise<Array<[*, *]>>} */
-  const getStoreSnapshot = async name =>
-    EV.vat('bootstrap').snapshotStore(await EV(powerStore).get(name));
-
-  const contractKits = await getStoreSnapshot('contractKits');
-  const psmKit = await getStoreSnapshot('psmKit');
-  const governedContractKits = await getStoreSnapshot('governedContractKits');
-  const vatStore = await getStoreSnapshot('vatStore');
-
-  /**
-   * Map refs to objects and find a vat containing one of them.
-   *
-   * @param {Record<string, unknown>} refs
-   * @param {string[]} exclude don't report hits from these vatIDs
-   */
-  const findVat = async (refs, exclude = [zoeVat]) => {
-    const mapped = {};
-    for await (const [prop, presence] of Object.entries(refs)) {
-      if (!presence) {
-        continue;
-      }
-      const obj = await EV.rawBoot.awaitVatObject({
-        presence,
-        rawOutput: true,
-      });
-      mapped[prop] = obj;
-    }
-    for (const obj of Object.values(mapped)) {
-      const [_k, vatID, _oid] = kernelTable.find(
-        row => row[0] === obj.getKref() && row[2].startsWith('o+'),
-      );
-      if (!exclude.includes(vatID)) {
-        return { vatID, mapped };
-      }
-    }
-    console.warn(`no vat found for`, refs);
-    return { vatID: undefined, mapped: {} };
-  };
-
-  for await (const [_instance, kit] of [
-    ...contractKits,
-    ...governedContractKits,
-    ...psmKit,
-  ]) {
-    t.truthy(kit.adminFacet || kit.psmAdminFacet, kit.label);
-    const { creatorFacet, publicFacet, psmCreatorFacet } = kit;
-    const { vatID, mapped } = await findVat({
-      creatorFacet,
-      publicFacet,
-      psmCreatorFacet,
-    });
-    console.log(
-      'kit',
-      { ...kit, ...mapped },
-      'has adminFacet of contract vat:',
-      vatID,
-      vatNames[vatID],
-    );
-    delete todo[vatID];
-  }
-
-  for await (const [_instance, kit] of [...governedContractKits, ...psmKit]) {
-    t.truthy(kit.adminFacet || kit.psmAdminFacet, kit.label);
-    const { governorCreatorFacet, psmGovernorCreatorFacet } = kit;
-    const { vatID, mapped } = await findVat({
-      governorCreatorFacet,
-      psmGovernorCreatorFacet,
-    });
-    console.log(
-      'kit',
-      { ...kit, ...mapped },
-      'has adminFacet of contract governor vat:',
-      vatID,
-      vatNames[vatID],
-    );
-    delete todo[vatID];
-  }
-
-  for await (const [name, info] of vatStore) {
-    t.truthy(info.adminNode, name);
-    const { root } = info;
-    const { vatID, mapped } = await findVat({ root }, swingsetVats);
-    if (vatID) {
-      console.log(
-        'kit',
-        { name, ...info, ...mapped },
-        'has adminNode of non-contract vat:',
-        vatID,
-        vatNames[vatID],
-      );
-      delete todo[vatID];
-    }
-  }
-
-  t.deepEqual(todo, {});
 });
